% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/element_line_outline.R
\name{element_line_outline}
\alias{element_line_outline}
\title{Outlined line theme element}
\usage{
element_line_outline(
  colour = NULL,
  linewidth = NULL,
  linetype = NULL,
  lineend = NULL,
  color = NULL,
  arrow = NULL,
  inherit.blank = FALSE,
  stroke_colour = NULL,
  stroke_color = NULL,
  stroke_linewidth = NULL
)
}
\arguments{
\item{colour, color}{Line/border colour. Color is an alias for colour.}

\item{linetype}{Line type. An integer (0:8), a name (blank, solid,
dashed, dotted, dotdash, longdash, twodash), or a string with
an even number (up to eight) of hexadecimal digits which give the
lengths in consecutive positions in the string.}

\item{lineend}{Line end Line end style (round, butt, square)}

\item{arrow}{Arrow specification, as created by \code{\link[grid:arrow]{grid::arrow()}}}

\item{inherit.blank}{Should this element inherit the existence of an
\code{element_blank} among its parents? If \code{TRUE} the existence of
a blank element among its parents will cause this element to be blank as
well. If \code{FALSE} any blank parent element will be ignored when
calculating final element state.}

\item{stroke_colour, stroke_color}{Outline colour. \code{stroke_color} is an alias
for \code{stroke_colour}.}

\item{stroke_linewidth}{Outline size in millimetres.}
}
\value{
An <element_line_outline> object.
}
\description{
Outlined line theme element
}
