% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/element_rect_outline.R
\name{element_rect_outline}
\alias{element_rect_outline}
\title{Outlined rectangle theme element}
\usage{
element_rect_outline(
  fill = NULL,
  colour = NULL,
  linewidth = NULL,
  linetype = NULL,
  color = NULL,
  stroke_colour = NULL,
  stroke_color = NULL,
  stroke_linewidth = NULL,
  inherit.blank = FALSE
)
}
\arguments{
\item{fill}{Fill colour.}

\item{colour, color}{Line/border colour. Color is an alias for colour.}

\item{linewidth}{Line/border size in mm.}

\item{linetype}{Line type. An integer (0:8), a name (blank, solid,
dashed, dotted, dotdash, longdash, twodash), or a string with
an even number (up to eight) of hexadecimal digits which give the
lengths in consecutive positions in the string.}

\item{stroke_colour, stroke_color}{Outline colour. \code{stroke_color} is an alias
for \code{stroke_colour}.}

\item{stroke_linewidth}{Outline size in millimetres.}

\item{inherit.blank}{Should this element inherit the existence of an
\code{element_blank} among its parents? If \code{TRUE} the existence of
a blank element among its parents will cause this element to be blank as
well. If \code{FALSE} any blank parent element will be ignored when
calculating final element state.}
}
\value{
An \verb{<element_rect_outline>} object.
}
\description{
The element for drawing outlined rectangles. Can be given as some arguments
in the \code{\link[ggplot2:theme]{ggplot2::theme()}} function.
}
\examples{
ggplot(mpg, aes(displ, hwy, colour = factor(cyl))) +
  geom_point() +
  theme(
    legend.key = element_rect_outline(
      colour = "green", stroke_color = "blue"
    )
  )
}
