% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/element_text_outline.R
\name{element_text_outline}
\alias{element_text_outline}
\title{Outlined text theme element}
\usage{
element_text_outline(
  family = NULL,
  face = NULL,
  colour = NULL,
  size = NULL,
  hjust = NULL,
  vjust = NULL,
  angle = NULL,
  lineheight = NULL,
  color = NULL,
  margin = NULL,
  inherit.blank = FALSE,
  stroke_colour = NULL,
  stroke_color = NULL,
  stroke_linewidth = NULL
)
}
\arguments{
\item{family}{Font family}

\item{face}{Font face ("plain", "italic", "bold", "bold.italic")}

\item{colour, color}{Line/border colour. Color is an alias for colour.}

\item{size}{Line/border size in mm; text size in pts.}

\item{hjust}{Horizontal justification (in \eqn{[0, 1]})}

\item{vjust}{Vertical justification (in \eqn{[0, 1]})}

\item{angle}{Angle (in \eqn{[0, 360]})}

\item{lineheight}{Line height}

\item{margin}{Margins around the text. See \code{\link[ggplot2:margin]{margin()}} for more
details. When creating a theme, the margins should be placed on the
side of the text facing towards the center of the plot.}

\item{inherit.blank}{Should this element inherit the existence of an
\code{element_blank} among its parents? If \code{TRUE} the existence of
a blank element among its parents will cause this element to be blank as
well. If \code{FALSE} any blank parent element will be ignored when
calculating final element state.}

\item{stroke_colour, stroke_color}{Outline colour. \code{stroke_color} is an alias
for \code{stroke_colour}.}

\item{stroke_linewidth}{Outline size in millimetres.}
}
\value{
An <element_text_outline> object.
}
\description{
The element for drawing outlined text. Can be given as some arguments in the
\code{\link[ggplot2:theme]{ggplot2::theme()}} function.
}
\examples{
ggplot(mpg, aes(displ, hwy)) +
  geom_point() +
  theme(
    axis.title.x = element_text_outline(
      colour = "white",
      stroke_linewidth = 0.5,
      stroke_colour = "red"
    )
  )
}
