% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geom_bar_outline.R
\name{geom_bar_outline}
\alias{geom_bar_outline}
\alias{geom_col_outline}
\alias{geom_histogram_outline}
\title{Outlined bar charts}
\usage{
geom_bar_outline(
  mapping = NULL,
  data = NULL,
  stat = "count",
  position = "stack",
  ...,
  just = 0.5,
  radius = NULL,
  by_group = TRUE,
  width = NULL,
  na.rm = FALSE,
  orientation = NA,
  show.legend = NA,
  inherit.aes = TRUE
)

geom_col_outline(..., stat = "identity")

geom_histogram_outline(..., stat = "bin")
}
\arguments{
\item{mapping}{Set of aesthetic mappings created by \code{\link[ggplot2:aes]{aes()}}. If specified and
\code{inherit.aes = TRUE} (the default), it is combined with the default mapping
at the top level of the plot. You must supply \code{mapping} if there is no plot
mapping.}

\item{data}{The data to be displayed in this layer. There are three
options:

If \code{NULL}, the default, the data is inherited from the plot
data as specified in the call to \code{\link[ggplot2:ggplot]{ggplot()}}.

A \code{data.frame}, or other object, will override the plot
data. All objects will be fortified to produce a data frame. See
\code{\link[ggplot2:fortify]{fortify()}} for which variables will be created.

A \code{function} will be called with a single argument,
the plot data. The return value must be a \code{data.frame}, and
will be used as the layer data. A \code{function} can be created
from a \code{formula} (e.g. \code{~ head(.x, 10)}).}

\item{position}{Position adjustment, either as a string naming the adjustment
(e.g. \code{"jitter"} to use \code{position_jitter}), or the result of a call to a
position adjustment function. Use the latter if you need to change the
settings of the adjustment.}

\item{...}{Other arguments passed on to \code{\link[ggplot2:layer]{layer()}}. These are
often aesthetics, used to set an aesthetic to a fixed value, like
\code{colour = "red"} or \code{size = 3}. They may also be parameters
to the paired geom/stat.}

\item{just}{Adjustment for column placement. Set to \code{0.5} by default, meaning
that columns will be centered about axis breaks. Set to \code{0} or \code{1} to place
columns to the left/right of axis breaks. Note that this argument may have
unintended behaviour when used with alternative positions, e.g.
\code{position_dodge()}.}

\item{radius}{A \code{\link[grid:unit]{grid::unit()}} object of length 1 setting the radius for
rounded corners.}

\item{by_group}{A \code{logical(1)} for setting groupwise outlines. If \code{TRUE}
(default), will draw outlines for groups. If \code{FALSE}, the collection of
bars as a whole get an outline.}

\item{width}{Bar width. By default, set to 90\% of the \code{\link[ggplot2:resolution]{resolution()}} of the
data.}

\item{na.rm}{If \code{FALSE}, the default, missing values are removed with
a warning. If \code{TRUE}, missing values are silently removed.}

\item{orientation}{The orientation of the layer. The default (\code{NA})
automatically determines the orientation from the aesthetic mapping. In the
rare event that this fails it can be given explicitly by setting \code{orientation}
to either \code{"x"} or \code{"y"}. See the \emph{Orientation} section for more detail.}

\item{show.legend}{logical. Should this layer be included in the legends?
\code{NA}, the default, includes if any aesthetics are mapped.
\code{FALSE} never includes, and \code{TRUE} always includes.
It can also be a named logical vector to finely select the aesthetics to
display.}

\item{inherit.aes}{If \code{FALSE}, overrides the default aesthetics,
rather than combining with them. This is most useful for helper functions
that define both data and aesthetics and shouldn't inherit behaviour from
the default plot specification, e.g. \code{\link[ggplot2:borders]{borders()}}.}
}
\value{
A \verb{<LayerInstance>} object that can be added to a plot.
}
\description{
Like \code{\link[ggplot2:geom_bar]{ggplot2::geom_bar()}} and \code{\link[ggplot2:geom_bar]{ggplot2::geom_col()}}, this draws a
barchart in the plot. However, the bars have an outline, that can make it
easier to discern groups.
}
\section{Aesthetics}{

\code{geom_bar_outline()} understands the following aesthetics (required aesthetics are in bold):
\itemize{
\item \strong{\code{x}}
\item \strong{\code{y}}
\item \code{alpha}
\item \code{colour}
\item \code{fill}
\item \code{group}
\item \code{linetype}
\item \code{linewidth}
\item \code{stroke_colour}
\item \code{stroke_linewidth}
}
Learn more about setting these aesthetics in \code{vignette("ggplot2-specs")}.
}

\examples{
# Bar doesn't need y-aesthetic
ggplot(mpg, aes(class, fill = factor(cyl))) +
  geom_bar_outline()

# Groups sharing outlines
ggplot(mpg, aes(class, fill = factor(cyl))) +
  geom_bar_outline(by_group = FALSE)

# Setup data
df <- data.frame(
  trt = c("a", "a", "b", "b", "c", "c"),
  outcome = c(2.3, 1.9, 3.2, 2.8, 2.0, 3.5),
  group = rep(c("A", "B"), 3)
)

# Dodged barchart sharing outlines
ggplot(df, aes(trt, outcome, fill = group)) +
  geom_col_outline(
    position = "dodge",
    by_group = FALSE
  )

# Histogram with outline
ggplot(iris, aes(Sepal.Length)) +
  geom_histogram_outline()

# Histogram with outline indicating totals
ggplot(iris, aes(Sepal.Length, fill = Species)) +
  geom_histogram_outline(by_group = FALSE)
}
