% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geom_polygon_outline.R
\name{geom_polygon_outline}
\alias{geom_polygon_outline}
\title{Outlined polygons}
\usage{
geom_polygon_outline(
  mapping = NULL,
  data = NULL,
  stat = "identity",
  position = "identity",
  rule = "evenodd",
  by_group = FALSE,
  ...,
  na.rm = FALSE,
  show.legend = NA,
  inherit.aes = TRUE
)
}
\arguments{
\item{mapping}{Set of aesthetic mappings created by \code{\link[ggplot2:aes]{aes()}}. If specified and
\code{inherit.aes = TRUE} (the default), it is combined with the default mapping
at the top level of the plot. You must supply \code{mapping} if there is no plot
mapping.}

\item{data}{The data to be displayed in this layer. There are three
options:

If \code{NULL}, the default, the data is inherited from the plot
data as specified in the call to \code{\link[ggplot2:ggplot]{ggplot()}}.

A \code{data.frame}, or other object, will override the plot
data. All objects will be fortified to produce a data frame. See
\code{\link[ggplot2:fortify]{fortify()}} for which variables will be created.

A \code{function} will be called with a single argument,
the plot data. The return value must be a \code{data.frame}, and
will be used as the layer data. A \code{function} can be created
from a \code{formula} (e.g. \code{~ head(.x, 10)}).}

\item{stat}{The statistical transformation to use on the data for this
layer, either as a \code{ggproto} \code{Geom} subclass or as a string naming the
stat stripped of the \code{stat_} prefix (e.g. \code{"count"} rather than
\code{"stat_count"})}

\item{position}{Position adjustment, either as a string naming the adjustment
(e.g. \code{"jitter"} to use \code{position_jitter}), or the result of a call to a
position adjustment function. Use the latter if you need to change the
settings of the adjustment.}

\item{rule}{Either \code{"evenodd"} or \code{"winding"}. If polygons with holes are
being drawn (using the \code{subgroup} aesthetic) this argument defines how the
hole coordinates are interpreted. See the examples in \code{\link[grid:grid.path]{grid::pathGrob()}} for
an explanation.}

\item{by_group}{A \code{logical(1)} for setting groupwise outlines. If \code{TRUE},
every polygon gets a separate outline. If \code{FALSE} (default), the collection
of polygons as a whole gets an outline.}

\item{...}{Other arguments passed on to \code{\link[ggplot2:layer]{layer()}}. These are
often aesthetics, used to set an aesthetic to a fixed value, like
\code{colour = "red"} or \code{size = 3}. They may also be parameters
to the paired geom/stat.}

\item{na.rm}{If \code{FALSE}, the default, missing values are removed with
a warning. If \code{TRUE}, missing values are silently removed.}

\item{show.legend}{logical. Should this layer be included in the legends?
\code{NA}, the default, includes if any aesthetics are mapped.
\code{FALSE} never includes, and \code{TRUE} always includes.
It can also be a named logical vector to finely select the aesthetics to
display.}

\item{inherit.aes}{If \code{FALSE}, overrides the default aesthetics,
rather than combining with them. This is most useful for helper functions
that define both data and aesthetics and shouldn't inherit behaviour from
the default plot specification, e.g. \code{\link[ggplot2:borders]{borders()}}.}
}
\value{
A \verb{<LayerInstance>} object that can be added to a plot.
}
\description{
Like \code{\link[ggplot2:geom_polygon]{ggplot2::geom_polygon()}}, this draws polygons in the plot. However,
the polygons have an outline, that can make it easier to discern groups.
}
\section{Aesthetics}{

\code{geom_polygon_outline()} understands the following aesthetics (required aesthetics are in bold):
\itemize{
\item \strong{\code{x}}
\item \strong{\code{y}}
\item \code{alpha}
\item \code{colour}
\item \code{fill}
\item \code{group}
\item \code{linetype}
\item \code{linewidth}
\item \code{stroke_colour}
\item \code{stroke_linewidth}
\item \code{subgruop}
}
Learn more about setting these aesthetics in \code{vignette("ggplot2-specs")}.
}

\examples{
# Setup some data
df <- data.frame(
  x = c(2, 6, 6, 2, 2.5, 5.5, 5.5, 2.5,
        8, 4, 4, 8, 7.5, 4.5, 4.5, 7.5),
  y = c(2, 2, 6, 6, 2.5, 2.5, 5.5, 5.5,
        8, 8, 4, 4, 7.5, 7.5, 4.5, 4.5),
  sub_id = rep(1:4, each = 4),
  id = rep(1:2, each = 8)
)

# Setup plot
p <- ggplot(df, aes(x, y, group = id, subgroup = sub_id, fill = factor(id)))

# Plot with polygon containing holes, sharing outlines
p + geom_polygon_outline()

# Outline for every polygon
p + geom_polygon_outline(by_group = TRUE)
}
