% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geom_rect_outline.R
\name{geom_rect_outline}
\alias{geom_rect_outline}
\alias{geom_tile_outline}
\title{Outlined rectangles}
\usage{
geom_rect_outline(
  mapping = NULL,
  data = NULL,
  stat = "identity",
  position = "identity",
  by_group = FALSE,
  ...,
  linejoin = "mitre",
  radius = NULL,
  na.rm = FALSE,
  show.legend = NA,
  inherit.aes = TRUE
)

geom_tile_outline(
  mapping = NULL,
  data = NULL,
  stat = "identity",
  position = "identity",
  by_group = FALSE,
  ...,
  linejoin = "mitre",
  radius = NULL,
  na.rm = FALSE,
  show.legend = NA,
  inherit.aes = TRUE
)
}
\arguments{
\item{mapping}{Set of aesthetic mappings created by \code{\link[ggplot2:aes]{aes()}}. If specified and
\code{inherit.aes = TRUE} (the default), it is combined with the default mapping
at the top level of the plot. You must supply \code{mapping} if there is no plot
mapping.}

\item{data}{The data to be displayed in this layer. There are three
options:

If \code{NULL}, the default, the data is inherited from the plot
data as specified in the call to \code{\link[ggplot2:ggplot]{ggplot()}}.

A \code{data.frame}, or other object, will override the plot
data. All objects will be fortified to produce a data frame. See
\code{\link[ggplot2:fortify]{fortify()}} for which variables will be created.

A \code{function} will be called with a single argument,
the plot data. The return value must be a \code{data.frame}, and
will be used as the layer data. A \code{function} can be created
from a \code{formula} (e.g. \code{~ head(.x, 10)}).}

\item{stat}{The statistical transformation to use on the data for this
layer, either as a \code{ggproto} \code{Geom} subclass or as a string naming the
stat stripped of the \code{stat_} prefix (e.g. \code{"count"} rather than
\code{"stat_count"})}

\item{position}{Position adjustment, either as a string naming the adjustment
(e.g. \code{"jitter"} to use \code{position_jitter}), or the result of a call to a
position adjustment function. Use the latter if you need to change the
settings of the adjustment.}

\item{by_group}{A \code{logical(1)} for setting groupwise outlines. If \code{TRUE},
will draw outlines for individual groups. If \code{FALSE} (default), the
collection of rectangles as a whole gets an outline.}

\item{...}{Other arguments passed on to \code{\link[ggplot2:layer]{layer()}}. These are
often aesthetics, used to set an aesthetic to a fixed value, like
\code{colour = "red"} or \code{size = 3}. They may also be parameters
to the paired geom/stat.}

\item{linejoin}{Line join style (round, mitre, bevel).}

\item{radius}{A \code{\link[grid:unit]{grid::unit()}} object of length 1 setting the radius for
rounded corners.}

\item{na.rm}{If \code{FALSE}, the default, missing values are removed with
a warning. If \code{TRUE}, missing values are silently removed.}

\item{show.legend}{logical. Should this layer be included in the legends?
\code{NA}, the default, includes if any aesthetics are mapped.
\code{FALSE} never includes, and \code{TRUE} always includes.
It can also be a named logical vector to finely select the aesthetics to
display.}

\item{inherit.aes}{If \code{FALSE}, overrides the default aesthetics,
rather than combining with them. This is most useful for helper functions
that define both data and aesthetics and shouldn't inherit behaviour from
the default plot specification, e.g. \code{\link[ggplot2:borders]{borders()}}.}
}
\value{
A \verb{<LayerInstance>} object that can be added to a plot.
}
\description{
Like \code{\link[ggplot2:geom_tile]{ggplot2::geom_rect()}} and \code{\link[ggplot2:geom_tile]{ggplot2::geom_tile()}}, this draws
rectangles in the plot. However, the rectangles have an outline, that can
make it easier to discern groups.
}
\section{Aesthetics}{

\code{geom_rect_outline()} understands the following aesthetics (required aesthetics are in bold):
\itemize{
\item \strong{\code{xmin}}
\item \strong{\code{xmax}}
\item \strong{\code{ymin}}
\item \strong{\code{ymax}}
\item \code{alpha}
\item \code{colour}
\item \code{fill}
\item \code{group}
\item \code{linetype}
\item \code{linewidth}
\item \code{stroke_colour}
\item \code{stroke_linewidth}
}
Learn more about setting these aesthetics in \code{vignette("ggplot2-specs")}.


\code{geom_tile_outline()} understands the following aesthetics (required aesthetics are in bold):
\itemize{
\item \strong{\code{x}}
\item \strong{\code{y}}
\item \code{alpha}
\item \code{colour}
\item \code{fill}
\item \code{group}
\item \code{height}
\item \code{linetype}
\item \code{linewidth}
\item \code{stroke_colour}
\item \code{stroke_linewidth}
\item \code{width}
}
Learn more about setting these aesthetics in \code{vignette("ggplot2-specs")}.
}

\examples{
p <- ggplot(tetris, aes(x, y, fill = type)) +
  coord_equal()

# Outline the collection of tiles
p + geom_tile_outline(colour = "black")

# Outline groups of tiles
p + geom_tile_outline(by_group = TRUE, colour = "black")

# Round corners of tiles
p + geom_tile_outline(radius = unit(2, "mm"), by_group = TRUE)

# Rectangles
df <- data.frame(
  xmin = c(0, 4, 6, 8,  10, 0, 4, 6, 8,  10),
  xmax = c(4, 6, 8, 10, 14, 4, 6, 8, 10, 14),
  ymin = rep(1:2, each = 5),
  ymax = rep(2:3, each = 5),
  z = factor(rep(1:5, each = 2))
)

ggplot(df, aes()) +
  geom_rect_outline(aes(fill = z), colour = "grey50")

ggplot(df, aes(xmin = xmin, xmax = xmax, ymin = ymin, ymax = ymax)) +
  geom_rect_outline(aes(fill = z), colour = "grey50", by_group = TRUE)
}
