% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geom_rect_outline.R, R/geom_bar_outline.R,
%   R/geom_line_outline.R, R/geom_points_outline.R, R/geom_polygon_outline.R,
%   R/geom_text_outline.R, R/ggoutlines-package.R
\docType{data}
\name{GeomRectOutline}
\alias{GeomRectOutline}
\alias{GeomTileOutline}
\alias{GeomBarOutline}
\alias{GeomPathOutline}
\alias{GeomLineOutline}
\alias{GeomPointOutline}
\alias{GeomPolygonOutline}
\alias{GeomTextOutline}
\alias{ggoutlines-ggproto}
\title{ggproto objects in ggoutlines}
\description{
The ggoutlines packages makes use of the ggproto class extension mechanism
of ggplot2. They ordinarily shouldn't be of any concern to the regular user,
as their access is implemented through the \verb{geom_*()} family of functions.
For developers, you can build your own ggproto objects inheriting from
ggoutlines' ggproto objects as usual, and as indicated in the examples.
}
\examples{
# Inheriting from ggoutlines' classes
MyOutlinePath <- ggproto(
  "MyOutlinePath", GeomPathOutline #, ...
)
}
\keyword{datasets}
