% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggoutlines-package.R
\docType{package}
\name{ggoutlines-package}
\alias{ggoutlines}
\alias{ggoutlines-package}
\title{ggoutlines: What the Package Does (One Line, Title Case)}
\description{
What the package does (one paragraph).
}
\author{
\strong{Maintainer}: Teun van den Brand \email{tahvdbrand@gmail.com} (\href{https://orcid.org/0000-0002-9335-7468}{ORCID})

}
\keyword{internal}
