% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grob_outline_line.R
\name{grob_outline_line}
\alias{grob_outline_line}
\title{Outlined line grob}
\usage{
grob_outline_line(
  x = unit(c(0, 1), "npc"),
  y = unit(c(0, 1), "npc"),
  stroke_lwd = 1,
  stroke_col = "black",
  id = NULL,
  id.lengths = NULL,
  super_id = NULL,
  default.units = "npc",
  arrow = NULL,
  name = NULL,
  gp = gpar(),
  vp = NULL
)
}
\arguments{
\item{x}{A numeric vector or unit object specifying x-values.}

\item{y}{A numeric vector or unit object specifying y-values.}

\item{stroke_lwd}{Line width of the outline.}

\item{stroke_col}{Colour of the outline.}

\item{id}{A numeric vector used to separate locations in \code{x} and
    \code{y} into multiple lines.  All locations with the same
    \code{id} belong to the same line.}

\item{id.lengths}{A numeric vector used to separate locations in \code{x} and
    \code{y} into multiple lines.  Specifies consecutive blocks of
    locations which make up separate lines.}

\item{super_id}{Identifiers of super-groups that share an outline.}

\item{default.units}{A string indicating the default units to use
    if \code{x} or \code{y} are only given as numeric vectors.}

\item{arrow}{A list describing arrow heads to place at either end
    of the line, as produced by the \code{arrow} function.}

\item{name}{ A character identifier. }

\item{gp}{An object of class \code{"gpar"}, typically the output
    from a call to the function \code{\link[grid]{gpar}}.  This is basically
    a list of graphical parameter settings.}

\item{vp}{A Grid viewport object (or NULL).}
}
\value{
A \link[grid:grid.grob]{grid::grobTree} object
}
\description{
Creates a graphical object consisting of outlined lines.
}
\examples{
require(grid)
grid.newpage()
grid.draw(grob_outline_line(
  gp = gpar(col = "limegreen", lwd = 5),
  stroke_col = "forestgreen",
  stroke_lwd = 5
))
}
