% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grob_outline_points.R
\name{grob_outline_points}
\alias{grob_outline_points}
\title{Outlined points grob}
\usage{
grob_outline_points(
  x = stats::runif(10),
  y = stats::runif(10),
  pch = 1,
  id = NULL,
  stroke_col = "black",
  stroke_lwd = 1,
  default.units = "native",
  name = NULL,
  gp = gpar(),
  vp = NULL
)
}
\arguments{
\item{x}{numeric vector or unit object specifying x-values.}

\item{y}{numeric vector or unit object specifying y-values.}

\item{pch}{numeric or character vector indicating what sort of
    plotting symbol to use.  See \code{\link{points}} for the
    interpretation of these values, and note \code{fill} below.}

\item{id}{Optional grouping identifier.}

\item{stroke_col}{Colour for the outline.}

\item{stroke_lwd}{Line width for the outline.}

\item{default.units}{string indicating the default units to use
    if \code{x} or \code{y} are only given as numeric vectors.}

\item{name}{character identifier. }

\item{gp}{an \R object of class \code{"gpar"}, typically the output
    from a call to the function \code{\link[grid]{gpar}}.  This is basically
    a list of graphical parameter settings; note that \code{fill} (and
    not \code{bg} as in package \pkg{graphics} \code{\link{points}}) is
    used to \dQuote{fill}, i.e., color the background of symbols with
    \code{pch = 21:25}.}

\item{vp}{A Grid viewport object (or NULL).}
}
\value{
A \link[grid:grid.grob]{grid::grobTree} object.
}
\description{
Creates a graphical object consisting of outlined points.
}
\examples{
require(grid)
grid.newpage()
grid.draw(grob_outline_points(
  x = rep(seq(0.1, 0.9, length.out = 5), 5),
  y = rep(seq(0.9, 0.1, length.out = 5), each = 5),
  gp = gpar(col = "red", fill = "grey50"),
  default.units = "npc",
  pch = 1:25,
  stroke_lwd = 1,
  stroke_col = "blue"
))
}
