% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grob_outline_polygon.R
\name{grob_outline_polygon}
\alias{grob_outline_polygon}
\title{Outlined polygon grob}
\usage{
grob_outline_polygon(
  x,
  y,
  id = NULL,
  id.lengths = NULL,
  pathId = NULL,
  pathId.lengths = NULL,
  super_id = NULL,
  rule = "winding",
  stroke_col = "black",
  stroke_lwd = 1,
  default.units = "npc",
  name = NULL,
  gp = gpar(),
  vp = NULL
)
}
\arguments{
\item{x}{A numeric vector or unit object specifying x-locations.}

\item{y}{A numeric vector or unit object specifying y-locations.}

\item{id}{A numeric vector used to separate locations in \code{x} and
    \code{y} into sub-paths.  All locations with the same
    \code{id} belong to the same sub-path.}

\item{id.lengths}{A numeric vector used to separate locations in \code{x} and
    \code{y} into sub-paths.  Specifies consecutive blocks of
    locations which make up separate sub-paths.}

\item{pathId}{A numeric vector used to separate locations in \code{x} and
    \code{y} into distinct paths.  All locations with the same
    \code{pathId} belong to the same path.}

\item{pathId.lengths}{A numeric vector used to separate locations in \code{x} 
    and \code{y} into paths.  Specifies consecutive blocks of locations which 
    make up separate paths.}

\item{super_id}{Identifiers of super-groups that share an outline}

\item{rule}{A character value specifying the fill rule: either
    \code{"winding"} or \code{"evenodd"}.}

\item{stroke_col}{Colour of the outline}

\item{stroke_lwd}{Line width of the outline}

\item{default.units}{A string indicating the default units to use
    if \code{x} or \code{y}
    are only given as numeric vectors.}

\item{name}{ A character identifier. }

\item{gp}{An object of class \code{"gpar"}, typically the output
    from a call to the function \code{\link[grid]{gpar}}.  This is basically
    a list of graphical parameter settings.}

\item{vp}{A Grid viewport object (or NULL).}
}
\value{
A \link[grid:grid.grob]{grid::grobTree} object
}
\description{
Creates a graphical object consisting of outlined polygons.
}
\examples{
require(grid)
grid.newpage()
grid.draw(grob_outline_polygon(
  x = c(0.2, 0.6, 0.6, 0.2, 0.25, 0.55, 0.55, 0.25,
        0.8, 0.4, 0.4, 0.8, 0.75, 0.45, 0.45, 0.75),
  y = c(0.2, 0.2, 0.6, 0.6, 0.25, 0.25, 0.55, 0.55,
        0.8, 0.8, 0.4, 0.4, 0.75, 0.75, 0.45, 0.45),
  id = rep(1:4, each = 4),
  pathId = rep(1:2, each = 8),
  gp = gpar(fill = c("dodgerblue", "tomato"), lwd = 2, col = "black"),
  stroke_col = "white",
  stroke_lwd = 5,
  rule = "evenodd"
))
}
