% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grob_outline_rect.R
\name{grob_outline_rect}
\alias{grob_outline_rect}
\title{Outlined rectangle grob}
\usage{
grob_outline_rect(
  x = unit(0.5, "npc"),
  y = unit(0.5, "npc"),
  width = unit(1, "npc"),
  height = unit(1, "npc"),
  just = "centre",
  hjust = NULL,
  vjust = NULL,
  id = NULL,
  r = NULL,
  stroke_col = "black",
  stroke_lwd = 1,
  default.units = "npc",
  name = NULL,
  gp = gpar(),
  vp = NULL
)
}
\arguments{
\item{x}{A numeric vector or unit object specifying x-location.}

\item{y}{A numeric vector or unit object specifying y-location.}

\item{width}{A numeric vector or unit object specifying width.}

\item{height}{A numeric vector or unit object specifying height.}

\item{just}{The justification of the rectangle
    relative to its (x, y) location.  If there are two values, the first
    value specifies horizontal justification and the second value specifies
    vertical justification.  Possible string values are: \code{"left"},
    \code{"right"}, \code{"centre"}, \code{"center"}, \code{"bottom"},
    and \code{"top"}.  For numeric values, 0 means left alignment
    and 1 means right alignment.
  }

\item{hjust}{A numeric vector specifying horizontal justification.
    If specified, overrides the \code{just} setting.}

\item{vjust}{A numeric vector specifying vertical justification.
    If specified, overrides the \code{just} setting.}

\item{id}{Optional grouping identifier.}

\item{r}{The radius of optional rounded corners.}

\item{stroke_col}{Colour for the outline.}

\item{stroke_lwd}{Line width for the outline.}

\item{default.units}{A string indicating the default units to use
    if \code{x}, \code{y}, \code{width}, or \code{height}
    are only given as numeric vectors.}

\item{name}{ A character identifier. }

\item{gp}{An object of class \code{"gpar"}, typically the output
    from a call to the function \code{\link[grid]{gpar}}.  This is basically
    a list of graphical parameter settings.}

\item{vp}{A Grid viewport object (or NULL).}
}
\value{
A \link[grid:grid.grob]{grid::gTree} object.
}
\description{
Outlined rectangle grob
}
\examples{
grid.newpage()
grid.draw(grob_outline_rect(
  x = c(0.33, 0.66),
  y = c(0.33, 0.66),
  width = 0.5, height = 0.5,
  stroke_lwd = 5,
  stroke_col = 3:2
))
}
