% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grob_outline_text.R
\name{grob_outline_text}
\alias{grob_outline_text}
\title{Outlined text grob}
\usage{
grob_outline_text(
  label,
  x = unit(0.5, "npc"),
  y = unit(0.5, "npc"),
  stroke_lwd = 1,
  stroke_col = "black",
  stroke_linejoin = "round",
  stroke_lty = 1,
  super_id = NULL,
  just = "centre",
  hjust = NULL,
  vjust = NULL,
  rot = 0,
  check.overlap = FALSE,
  default.units = "npc",
  name = NULL,
  gp = gpar(),
  vp = NULL
)
}
\arguments{
\item{label}{A character or \link{expression} vector.  Other
    objects are coerced by \code{\link{as.graphicsAnnot}}.}

\item{x}{A numeric vector or unit object specifying x-values.}

\item{y}{A numeric vector or unit object specifying y-values.}

\item{stroke_lwd}{Line width for the outline.}

\item{stroke_col}{Colour for the outline.}

\item{stroke_linejoin}{Line join style for the outline.}

\item{stroke_lty}{Line type for the outline.}

\item{just}{The justification of the text
    relative to its (x, y) location.  If there are two values, the first
    value specifies horizontal justification and the second value specifies
    vertical justification.  Possible string values are: \code{"left"},
    \code{"right"}, \code{"centre"}, \code{"center"}, \code{"bottom"},
    and \code{"top"}.  For numeric values, 0 means left (bottom) alignment
    and 1 means right (top) alignment.
  }

\item{hjust}{A numeric vector specifying horizontal justification.
    If specified, overrides the \code{just} setting.}

\item{vjust}{A numeric vector specifying vertical justification.
    If specified, overrides the \code{just} setting.}

\item{rot}{The angle to rotate the text.}

\item{check.overlap}{A logical value to indicate whether to
    check for and omit overlapping text.}

\item{default.units}{A string indicating the default units to use
    if \code{x} or \code{y}
    are only given as numeric vectors.}

\item{name}{ A character identifier. }

\item{gp}{An object of class \code{"gpar"}, typically the output
    from a call to the function \code{\link[grid]{gpar}}.  This is basically
    a list of graphical parameter settings.}

\item{vp}{A Grid viewport object (or NULL).}
}
\value{
A \link[grid:grid.grob]{grid::gTree} object.
}
\description{
Creates a graphical object consisting of outlined text.
}
\examples{
require(grid)
grid.newpage()
grid.draw(grob_outline_text(
  label = c("Hello there,", "General Kenobi"),
  y = c(0.66, 0.33),
  gp = gpar(col = c("white", "red"), fontsize = 32),
  stroke_col = c("grey50", "black"), stroke_lwd = 3
))
}
