% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggoutlines-package.R
\docType{package}
\name{ggoutlines-package}
\alias{ggoutlines}
\alias{ggoutlines-package}
\title{ggoutlines: Outlines for 'ggplot2'}
\description{
A 'ggplot2' extension that provides extra geometry layers (geoms) that place outlines around some well-known geoms. Alongside the geoms, it also provides outlined graphical primitives to work in the 'grid' package.
}
\author{
\strong{Maintainer}: Teun van den Brand \email{tahvdbrand@gmail.com} (\href{https://orcid.org/0000-0002-9335-7468}{ORCID})

}
\keyword{internal}
